Implement collection service with ownership checks and toggle functionality
